Review with respect to reduce complexity
The copy and the assignment constructor are deleted. The default constructor now makes use of constructor delegation and thus reduced complexity.

The definition of two member functions was moved to the source file.

Since the class SimulationServer knows the input and output values for each simulation, we can omit the values parameter in function sendOutputValues.